refactor(client): [CLI-035] :recycle: Update clean code

+++ client/core/src/main/java/fr/esgi/pa/alliodesk/core/manager/TokenManager.java
@@ -30,6 +30,7 @@ public class TokenManager {
 
     public static void deleteTokenFile(){
         File file = new File("Token.txt");
+        if(file.exists())
         file.delete();
     }
 }
